feat: Convert declared licenses from meta-doubleopen to SDPX expressions
meta-doubleopen saves the declared licenses of recipes to the SPDX documents it
creates. The licenses may not be correct SPDX license expressions, so convert
them to be correct when populating the document from Fossology.

This functionality would better fit in the meta layer (as it's done in the SPDX
creation in Poky), but older versions of Poky would require bigger work to do
this.

Signed-off-by: Mikko Murto <mikko.murto@hhpartners.fi>